refactor: move CLI files to dedicated cli directory
